+++ Windal Renderer/include/core/entities/cameraentity.h
@@ -8,7 +8,7 @@ public:
 	~CameraEntity();
 
 protected:
-	virtual void UpdateSelf(float delta) override;
+	virtual void UpdateSelf(double delta) override;
 	virtual void RenderSelf(RenderServer& renderServer) override;
 	virtual void RenderImguiSelf() override;
 };
